Add Discover method and test node address
Probably will add a dependency because we need to parse dynamic JSON. Or I will figure something out. Discover method, at least with official node, responses with 1.5MB JSON that have basically all the information about the node.

While you can generate a struct from the JSON response - I think this is a bad idea because there's a lot of dynamic fields in schema. This is an absolute hell.